readjusted rollConfig optionals and hbs - introduced rollConfigHelpers.mjs
