+++ templates/subapplications/optionalBonuses.hbs
@@ -1,21 +1,36 @@
 <div class="babonus-optionals">
   {{#each data as |bab|}}
+
+  {{#unless bab.consumes}}
   <div class="optional">
     <div class="bonus-text">
       <span class="name">{{bab.name}}.</span>
-      <span class="description">{{bab.desc}}</span>
+      <span class="description">{{{bab.desc}}}</span>
     </div>
     <a class="add" data-bonus="{{bab.bonus}}">
       <i class="fa-solid fa-plus"></i>
     </a>
-    {{#if bab.consumes}}
-    <div class="consumption" data-min="{{bab.min}}" data-uuid="{{bab.uuid}}" data-type="{{bab.type}}">
-      {{#if bab.scales}}
+  </div>
+  {{/unless}}
+
+  {{#if bab.consumes}}
+  <div class="optional" data-min="{{bab.min}}" data-uuid="{{bab.uuid}}" data-type="{{bab.type}}">
+    <div class="bonus-text">
+      <span class="name">{{bab.name}}.</span>
+      <span class="description">{{{bab.desc}}}</span>
+    </div>
+    <a class="add" data-bonus="{{bab.bonus}}">
+      <i class="fa-solid fa-plus"></i>
+    </a>
+
+    {{#if bab.scales}}
+    <div class="consumption">
       <label>{{localize "BABONUS.CONSUMPTION_APP.LABEL"}}</label>
       <select>{{{bab.options}}}</select>
-      {{/if}}
     </div>
     {{/if}}
   </div>
+  {{/if}}
+
   {{/each}}
 </div>
